Moved quelware-derived files.
